requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
fake-useragent==0.1.11
gunicorn==20.0.4
idna==2.9
lxml==4.5.0
parse==1.15.0
Pillow==7.1.1
pyee==7.0.1
pyppeteer==0.0.25
pyquery==1.4.1
requests==2.23.0
requests-html==0.10.0
six==1.14.0
soupsieve==2.0
tqdm==4.44.1
urllib3==1.25.8
vk-api==11.8.0
w3lib==1.21.0
websockets==8.1
